Move `NullStore` to its own file
