+++ EssentialApp/EssentialApp/NullStore.swift
@@ -5,4 +5,28 @@
 //  Created by Julius on 11/07/2024.
 //
 
+import os
 import Foundation
+import essential_feed_case_study
+
+class NullStore: FeedStore & FeedImageDataStore {
+    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        completion(.success(()))
+    }
+    
+    func insert(_ feed: [essential_feed_case_study.LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        completion(.success(()))
+    }
+    
+    func retrieve(completion: @escaping RetrievalCompletion) {
+        completion(.success(.none))
+    }
+    
+    func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
+        completion(.success(.none))
+    }
+    
+    func insert(_ data: Data, for url: URL, completion: @escaping (InsertionResult) -> Void) {
+        completion(.success(()))
+    }
+}
